fix: a bunch of bugs regarding incorrect startup order and LSP not triggering
