+++ lua/plugins/configs/lsp-signature.lua
@@ -1,0 +1,4 @@
+return function()
+    -- Signature help
+    require('lsp_signature').on_attach()
+end
